requirements.txt
requests
moviepy
pytrends
pandas
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
